fix: create multiple SNIs when same certificate is reused
As seen in #75, if multiple Host reference the same certificate, only
one host is added as SNI in Kong. This was a bug due to early exit when
a GET request for a certificate succeded.
There were also edge-cases around change certificate for a Host or
updating SNI name. This commit covers those.