feat(KongConsumer) delete consumers not present in k8s
Problem:
When `KongConsumer` custom resource is created, a consumer entity is
created in Kong. But when the custom resource is deleted in k8s, the
change is not propagated to Kong. This results in an odd and confusing
behavior(#58) where consumers can be created but never deleted in Kong via
ingress controller.

Solution:
Perform complete sync like we do for all other entities in Kong. This
means that after this change, consumers present in Kong but not present
in k8s will be deleted.
This reverses the past decision of allowing users to manage consumer entity in Kong via ingress controller and by other means outside the controller.

The long term solution will involve storing more metadata in Kong to
detect if an entity is created by the ingress controller or via another
mechanism.

Breaking change:
Consumers created in Kong directly will be deleted if the corresponding
custom resource (`KongConsumer`) is not present in k8s object store.

From #81