feat: introduce a Name parameter fo KongPlugin CRD
This is part of a patchset introducing support for global plugins in
Kong via the ingress controller.
This is a backwards compatible change.

This commit will have long-term effects in how plugins are configured
via Ingress and custom resources as explained below:

Currently, to apply a plugin to one has to add an annotation of the
form:
```
rate-limiting.plugin.konghq.com: plugin-conf
```
where `rate-limiting` is the name of the plugin that needs to be
created while `plugin-conf` is the plugin conf that should be used for
that plugin.

This is not as user-friendly as it should be, especially considering
that different kinds of plugins in Kong have different configuration
objects.

This allows us to keep configuration and plugin names together and in
future the annotations could be much more compact and easier to pass
around:
```
plugin.konghq.com: plugin-conf1,plugin-conf2
```