fix: rename name to plugin in KongPlugin resource
This is a fix for the issue introduced in 32ca70f.
`name` is already a property defined for every resource in k8s, which is
unique across all objects of the same kind.
The `name` from metav1.objectMeta needs to be accessible in other parts
of code to correctly read the object.

This commit renames the `name` to `plugin` field.