feat: introduce a helper method to parse `plugins.konghq.com` annotations
Currently, plugins in Kong are configured using the following
annotation: `key-auth.plugin.konghq.com: <name of KongPlugin CR>`.

This works fine but decouples the configuration of a plugin from the
name of the plugin to which that configuration should be applied to.
In practise, it rarely happens that plugin configs are reusable across
different types/kinds (key-auth, rate-limiting etc) of plugins.

This change introduces a method to parse an annotation of the form
`plugins.konghq.com`, which allow you to specify KongPlugin Custom
Resources(CR) as a CSV, which can be applied to an entity in Kong (route
or service or consumer).