feat: allow configuring KongPlugins using `plugins.konghq.com`
This change allows an end-user to configure a plugin on a route or a
service in Kong using a new annotaion: `plugins.konghq.com`.
This annotations should contain a CSV list of KongPlugin CR names which
should be applied.

This should make it easier for the end user to configure plugins on
Ingress resources and k8s Services.

The old way of configuration plugins:
`<plugin-name>.plugin.konghq.com: <KongPlugin CR name>`
is deprecated but still supported, giving users time to migrate to new
annotations. There will be a documentation commit made in future to
reflect the same.