feat(deploy) add initContainer to wait for DB state
Kong can startup only if the DB schema is up to date.
If Postgres is not up or the schema is not up to date, then Kong
will fail to start, resulting in Pod restarts.

This change will wait for Postgres to be up and migrations to finish
running, avoiding restarts and failures in Kubernetes.

There is currently no way to verify if DB state is at the correct
version and hence we resort to a hack of actualling starting Kong and
stopping it in the init container.

With Kong >= 0.15.0, Kong will have a CLI command to check if migrations
are up to date or not.