fix: add missing default value in Upstream
Kong 1.0 introduced support for HTTPS active healthchecks, which was
introduced in go-kong v0.4.0.
With this support, if the default values are not specified, deck detects
this as a drift in configuration and issues a PUT request for all
usptreams in every sync.
Supplying these values avoids the diff.

Since the field was introduced in Kong 1.0, this change
is compatible with Kong 1.0 and Kong 1.1.