hotfix: inject protocols in plugins for >= Kong 1.1
Kong 1.1 introduced `protocols` field in the Plugin entity.

Since this field is not present in Kong 1.0, if the field is injected by
the Ingress Controller, Kong 1.0 will error out as it is not aware of
existence of such a field.

This commit injects the Protocols field only if Kong's version is >=
1.1.0.

Future versions of Kong Enterprise will also add support for `protocols`
field in Plugin, and when this happens we will include support for that
version in the Ingress Controller.