feat: drop --default-backend-service CLI flag
The flag comes in from the era of NGINX ingress controller.

The default service is selected from the Ingress rules if any of the
Ingress rules contain a backend at the root level, meaning a service
which is the fallback is none of the Ingress rules actually match.

While this flag was present, it is not being used anywhere in the
codebase and will not be used in future.

This is a breaking change and users are advised to update their
YAML deployments to remove the flag, else the controller will fail to
boot up.