fix: always run leader election even when update-status is false
Previously, if status need not be synced, then leader election would
fail and ingress controller would fail to start.

This commit separates the logic out in a hacky way.
With DB-less, if status of Ingress resource is not to be updated, then
we don't even need to run leader-election.

In a future commit, leader election logic and sync logic can be
completely separated out, unlike currently, when on shutdown the sync
logic updates the configmap.

Fix #232
From #253