refactor: consolidate autogenerated k8s client into a single package
The autogenerated clients all belong to `configuration.konghq.com` API
group in k8s and should be contained inside a single package.

This simplifies the code that is using the package for interracting with
k8s API-server as only a single ClientSet is needed which is capable of
interracting with all of our custom resources.