fix: ensure multiple certs are created
Not assigning a certificate ID to the cert being added to deck's
target state results in only the most recently updated cert to be
persisted in Kong.

While we fix this issue by correcting the typo, we are not able to add a
unit or integration test at this point. This is being addressed with the
on-going work to refactor parser and store out of the controller logic
and also with integration tests and simpler API in deck for using deck
as a library.

Fix #285