refactor: separate out store implmentation
Store was a concept internal to the controller package.
This meant that the only way to actually write test was run the entire
Ingress Controller.

This commit refactors the store into it's own dedicated package and we
can inject a store into the Ingress Controller.
This makes it possible to create a fake store and write integration
tests which cover most of our translation logic using the parser and a
fake store.

Changelog:
- Events are no longer populated for the Ingress resouce in k8s.
- Separate store package with store implementation
- All the informers are now instantiated and controlled in the main
  package.