tests: add a Fake Store implementation
This is to be used to perform testing for the Parser package.

In future, if we would like to perform a complete integration test of
the Ingress controller along-with Kong, then it will be possible.

This abstracts away most of the k8s APIs that we are using (less leader
election and informer logic) for fetching data.