feat: add check_hash=1 to db-less /config API call
Kong 1.2 ships with a feature to avoid a redundant config reload if the
config has not changed at all.

The Ingress Controller is already doing a sha sum and checking if it
needs to make a call to the /config endpoint. This adds in another layer
of checking the same logic on the Kong side.