fix: do not error out due to a missing service or plugin
If an Ingress resource references a plugin which is not present in k8s's
store, ingress controller bails out and stops syncing any other
configuration updates as well.

This commit changes that and allows user to proceed in case where
Ingress is applied but the service is not present in the system.

The test cases for these are in works with a fake store implementation.

It is not ideal that all errors are logged and we never error out. An
error from, let's say, looking up an item from cache, should result in a
failure to sync, while an error of an item not existing can merely be
logged.
This is again in the works, where errors will be classified and metrics
around the errors exposed for alerting in Prometheus.