feat: introduce ingress.kubernetes.io/service-upstream annotation
Problems
--------

1. Kong Ingress Controller configures targets in Kong based on the
   endpoints of the corresponding Kubernetes service. This allows Kong
   to be used a load-balancer, but Kong then by-passes kube-proxy.
   There are use-cases where kube-proxy should handle load-balancing
   and Kong should send traffic to kube-proxy instead of pods.
   This was not possible.
2. In Kong, the `host` header sent to upstream is set to the IP
   value of the target picked for the request (unless `preserve_host`
   is set to true in the route selected for routing the request).
   This breaks routing in certain cases(see #330) when the upstream
   routes or responds based on the `host` header it receives in the
   request.

Solution
--------

Both of these problems are being addressed by populating the target of
the service with the DNS entry of the Kubernetes service i.e.
`service.namespace.svc`. This means , if the
`ingress.kubernetes.io/service-upstream` annotation is set to `true`,
then Kong will send the traffic to the ClusterIP of the service and also
set the host header to `service.namespace.svc`.
In a future release of Kong, Kong will support setting an arbitrary
`host` header value when sending a request upstream, which will allow
for further customization as needed.

See #330 #48