feat: support for live syncing of credentials
Credential sync for DB mode has now been moved to decK:
- multiple credentials of the same type can be provisioned
- any change to credentials will now be propagated to Kong #230

With this change, all syncing is now done via decK, providing for a
stable sync with least API calls.

There exists one caveat here, any change to password for basic-auth will
not result in a change in Kong because it is not possible to diff
passwords as Kong hashes them in the API response and DB. There will be
a follow up commit to work around this limitation.

For DB-less mode, the parsing becomes simpler but no functional change
should happen.

This change also lays the foundation for secret based credentials in future.