feat: sync basic-auth password changes in DB mode
With passwords hashed in Kong, decK can not detect a drift when the
password is changed.
This change syncs the password change to Kong.
Username changes result in re-creation of the credential in Kong and
they are taken care of by decK.

This change has no effect on the DB-less deployments.