feat: implement admission controller for configuration.konghq.com group
Users frequently stumble onto situation where a misconfigured plugin
stops the entire controller from functioning.

It is now possible to setup an admission validation webhook to prevent
users from mis-configuring plugins and consumers.

Documentation on how to configure the controller will be published soon
and this controller will include KongCredentials in the future.