feat(deploy) unify controller and proxy for DB mode
This commit brings a lot of enhancements in how Kong Ingress Controller
is deployed:
- Custom server block to expose health of Kong and Prometheus metrics is
  now common to DB and DB-less deployments.
- DB deployment now consists of only a single Deployment, with a Kong proxy
  container, and the Ingress Controller. The admin interface is
  available on the loopback interface. This makes the delta between the
  the two deployment modes (DB and DB-less) very small.
- Naming of various deployments and ports has been changed.
- By default, the replica is set to 2.

These changes also lay the ground for adding support for `kustomize`.
All the common manifests will be moved to a dedicated directory, and
DB and DB-less modes will be overlays on top of the common directory.