feat: ensure Kong Enterprise 0.36 compatiblity
Kong Enterprise 0.36 is based on Kong 1.2 core:
https://docs.konghq.com/enterprise/changelog/#036

`https_redirect_status_code` property was introduced in Kong 1.2 and
this needs to injected when Kong Enterprise is running.
This ensures that decK detects no change in state and fires no
Admin API requests while performing a full-sync during idle conditions
i.e. when there are no changes in the k8s cluster.