feat(cli) implement environment variable based configuration
In addition to CLI based configuration, environment variables can now be
used to configure the Ingress Controller.
All env vars are prefixed with `CONTROLLER_`, followed by the name of
the CLI flag (`-` converted to `_`).

The configuration is parsed into an intermediate struct and then later a
subset is passed down to the controller config. This is a change to make
split out controller and make it unit-testable in future, an effort
on-going since last few releases.