feat(cli) implement --kong-admin-filter-tag flag
The new flag can be used to tweak the tag that is used by the Ingress
controller to filter entities it manages and also tag entities it
creates.

This is configurable and should not be changed once the Ingress
Controller is deployed with a database. If the need be, then the
database should be truncated before changing this.