hotfix: rename credType to kongCredType for secret-based credential
Previously, `credType` key was used to lookup the type of kong
credential that was defined.

This key can potentially collide with other use-cases of Secret resource
in Kubernetes.

Hence, this key is renamed to `kongCredType`, which also provides good
context to the end user that this secret has something to do with a
credential inside Kong.