fix: populate algorithm of Upstream for Kong >= 1.3.0
The `algorithm` in the `upstream` entity defaults to `round-robin`.
This field was introduced in Kong 1.3.0.

If this field is not set, decK finds a diff between the current state
and target state and issues a PUT request to Kong, updating the upstream
in Kong's database.

This leads to a cache invalidation and rebuilt of upstreams in Kong,
causing performance degradation.

The code now injects the algorithm based on the version of Kong (or Kong
Enterprise). Kong Enterprise 1.3.0.0 will also pick this change
correctly.