chore: bump up pkg/errors to v0.9.1
To use `As()` function.
This is same as the new errors.As function introduced in Go 1.13.
We are using the library's function because we use it all over the
place.
We should refactor to remove thie library and use the stdlib functions
introduced in Go 1.13 that provide the same functionality.