feat(parser) add support for KongClusterPlugins
KongClusterPlugins allows user to have Plugin configuration at a
cluster-level and share the configuration across multiple workspaces.

This also helps users who want to have different personas for defining a
plugin configuration and using a plugin on a service or ingress as
access to KongClusterPlugin can be setup independent of other resources.

KongPlugin in the namespace takes precedence over KongClusterPlugin.

Global plugins for KongClusterPlugins are also supported.

Based on feedback, in future, we will deperecate global-level KongPlugin
support and instead only KongClusterPlugins will support the global
label. This will solve the problem of a user with access to a
single namespace manipulating proxy behavior for services in all namespaces.