fix(parser) handle same SNI across different Ingress rules
When the same SNI is associated with different Kubernetes Secrets.
When using Kong with a database, this results in the same SNI being
associated with a different certificate and the controller fails to
update the SNI in Kong.

This fix reduces the likelihood of this happening for some cases but
will not fix the problem for all cases.

See
https://github.com/hbagdi/deck/commit/890404c3012a620bf46b220705a54e338e5387b1
for the root cause of this issue.

Fix #510