feat: disable strip path by default to conform to Ingress spec
Ingress specification isn't clear but it is assumed that proxies do
transparent proxying, i.e. they don't alter or change the requests as
they are proxied to the Kubernetes service.

This is an unfortunate breaking change that will cause problems with
adoption of the new version. Strip path setting can be tweaked either
via KongIngress resource or via an annotation on the Ingress resource.