fix: correctly set knative status
Knative requires DomainInternal in the status to be populated as it is
used for in-cluster communication.

Also fixed a corner case which occurs in knative conformance test only
where no headers are added. Kong 500s if all fields in
config of request-transformer are empty. This is a bug in Kong but we
are working around it with this change.