feat: implement secret based ca-certificates
CA certificate is a core entity in Kong that enables storing ca
certificates in kong to be used for mtls verification of client
certificates.

This patch adds the ability to configure CA certificates in Kong for
both DB and DB-less modes.
All secrets containing the label `konghq.com/ca-cert: "true"` are
processed. The secrets should contain `id` and `cert` keys, which should
contain UUID and PEM-encoded certificate respectively.

Please note that secrets need to be labelled and not annotated. This is
intentional to keep search for the relevant secrets efficient.

The secret UID is not used and instead an explicit UUID is required to
support declarative configuration. One can't specify the UUID of
Kubernetes objects during create time, which means, the secret UUID will
change from run-to-run. References to ca certificates are resolvable
only via UUID, meaning the UUID needs to be constant.