feat: deprecate cassandra compatibility
Running the controller with cassandra has caused all sorts of pains for
users.

We already don't provide any tooling and documentation around
running ingress controller with kong and cassandra.
Helm chart removed easy cassandra mode a while ago, this change emits a
warning. In a future version, the controller will not startup if kong is
running with cassandra as it's data-store.

Based on the current statistics, there is a very small group of
users that will be affected by this change.