deploy: change externalTrafficPolicy to the default of `Cluster`
Originally, the policy was set to `Local` in order to preserve client IP
address out of the box. It turns out that doing so has been more harmful
and frustrating for most users.

Managed k8s offerings other than GKE don't respect this setting.
Especially, traffic routing from an NLB/ELB breaks down terribly if the
policy is changed to `Local`. Google's implementation has/had a bug
where TCP connections are throttled if the policy is `Local`.

Due to these issues, the service will use the default value of
`Cluster`.
This patch only removes the property rather than changing it to avoid
affecting any existing installations.

No installation should use these manifests as is, but rather create a
copy. We don't know how users rely on this so we are going to err on the
safer side.