deploy: drop custom-server-block and use status_listen
Until Kong 1.4, there didn't exist a way to health-check kong and expose
metrics without also exposing the admin api of kong.
In kong 1.4, a new server KONG_STATUS_LISTEN was added which provides
these capabilities.

This commit removes the reliance on the custom server block that was
provided via a ConfigMap.
With 0.9, the next version of controller, the minimum version of kong
should be 1.4, hence this breaking change.

The error logs have been modified as well to include proxy error logs as
well.