docs: specify key in --from-file parameter
Enterprise users frequently run into problems with the current way the
secret is created. The files have extensions which cause problem later
on in the deployment and this has taken up a lot of time in debugging
several support tickets.

The new format makes it clear and avoids problems when the file has an
extension or is present in a different directory.