fix: correctly find other pods when exiting
When the controller exits, it tries to search for other ingress
controller pods running in the cluster. It does so by looking up pods
which contain the same labels.

During an upgrade or any change to Deployment spec, the
pod-template-hash label value changes because ReplicaSet controller
needs to uniquely monitor the pods that are part of the previous version
and the new version. This leads to an incorrect logic where the
controller falsely assumes that there are no other pods running in the
cluster.

This commit fixes the logic for specific deployments but doesn't address
the problem for deployments of all types. This is discussed in more
detail in the connected issue.

Relates to #581