fix: do not require any jwt-auth fields
Per https://docs.konghq.com/hub/kong-inc/jwt/#create-a-jwt-credential
there are /no/ required fields for jwt-auth other that the associated
consumer. This is kinda true but not really: fields may be required
depending on the value of other fields, or may be required but generate
random values if none is presented.

In any case, the validation logic is complex enough that it makes more
sense to remove this from the webhook and let Kong handle it.

Fix #750