feat(store) handle Ingress v1 class
Update the Ingress v1 lister to handle both the legacy annotation used
in Ingress v1beta1 and the native Ingress class field introduced in
Ingress v1.

If the legacy annotation is present, the controller uses its value to
decide whether or not to process the Ingress, regardless of whether the
new native field is present. If the legacy annoation is absent, the
controller will use the native field. The decision logic itself is
unchanged: it continues to require an exact class match by default, and
can optionally also match Ingresses with no annotation or native class
if users override this default.