test(annotations+store) add Ingress v1 class tests
Test Ingress v1 class matching. In annotations, test that the match
behaviors are the same for native class matching as they are for ingress
class annotations. In store, test that class matching successfully
filters Ingresses with the correct class into the store and that
Ingresses with the correct class, but with an incorrect class
annotation, are not filtered in, because the class annotation takes
precedence.