feat(transl) disable automatic SNI handling
Disable automatic SNI criteria addition by commenting out the code that
adds it. This is a useful convenience feature, but would cause breakage
because of https://github.com/Kong/kong/issues/6425

Once that issue is fixed, we should re-enable this functionality, but
gate it with a flag, to avoid issues with older Kong versions.