Add SNI match criteria to routes (#863)
* feat(transl) add SNI match criteria to routes

Add SNI match criteria to Kong routes when the source Ingress has a TLS
rule with hostnames.

* feat(override) add SNI annotation and KongIngress

Add an Ingress annotation and KongIngress functionality for setting
route SNI match criteria. Although the controller adds SNI criteria
automatically for Ingresses that contain rules with hostnames and have
TLS information, there are some scenarios that require setting SNI
manually:

* You wish to strip SNI match criteria from the route, in order to
  support SNI-incapable clients. SNI annotation processing is unusual to
  accommodate this: providing a "konghq.com/snis: ''" annotation is
  distinct from not providing the annotation at all, and instructs the
  controller to strip SNI information that it'd normally add
  automatically.
* You use wildcard hostnames in your rules. Kong does not allow wildcard
  match criteria as of 2.1, and as such you must enumerate the specific
  SNI matches you want when using a wildcard hostname.
* You wish to use a completely different SNI match criteria that does
  not match your rule hostnames at all. This is highly atypical, and
  should only be done if you know exactly why you need to do it.

* feat(transl) disable automatic SNI handling

Disable automatic SNI criteria addition by commenting out the code that
adds it. This is a useful convenience feature, but would cause breakage
because of https://github.com/Kong/kong/issues/6425

Once that issue is fixed, we should re-enable this functionality, but
gate it with a flag, to avoid issues with older Kong versions.

* chore(deploy) update manifests

* pr(anns) use consistent pluralization

* fix(anns) use nil in place of empty string slice

* pr(parser) use consistent caps and label unsafe

Use SNI (rather than sni) throughout, and prefix which version is unsafe
rather than which version is safe.

* pr(parser) remove temporary commented code

* test(anns) expect nil slice for empty annotations

* test(transl) add E2E test for SNI annotation

* chore(deploy) update single manifests

* pr: add 404 test

* pr: fix magic number

* test(sni): add cases: 404 and server cert

Co-authored-by: Michał Flendrich <michal@flendrich.pro>