feat(*) disable KongClusterPlugins if unavailable
Check if KongClusterPlugins are available from the API Server. If not,
do not create their informer or store, and return errors from their
store functions.

Note: the Kubernetes API does not provide a means to retrieve a single
API resource, only all resources within a group. This retrieves the
entirety of "configuration.konghq.com/v1" and iterates through it to
check if KongClusterPlugin is included, and disables it if not. This is
a bit cumbersome, but should be fine, as the resource set is small and
we only need to check for it once at startup.