feat(transl) generate consistent route names
Instead of generating route names like namespace.ingress.XY, where X is
the rule index in the Ingress and Y is the path index within the rule,
generate route names like namespace.ingress.Z, where Z is the CRC32
checksum of the rule host and path (or port, for TCPIngress rules).

This guarantees that the order of rules within an Ingress cannot affect
the route name.

Fix #834