fix(kongstate): missing logger fields (#1014)
**What this PR does / why we need it**:
`log.WithFields(...).Errorf()`  emits a log entry with all the recorded fields.
However, `log.WithFields(...).Logger.Errorf()` emits a log entry too, but without these fields.

This PR fixes the logger invocation by preventing it from dropping the recorded fields, by switching from calling the latter `Errorf` to the former.

**Which issue this PR fixes**
fixes #1002

**Special notes for your reviewer**:
This logger invocation (and tens of others) includes the `error` in the message. With structured logging, the preferred way is to put an error in a field of its own. Filed #1013 to take care of that.