feat: ctrl test hack now capable of running legacy
This expands on the existing hack which runs the controller
for integration tests by making it capable of running the
legacy (v1.x) version of the controller as well. This is
needed in the interim between v1.x and v2.x so that all of
the integration tests run (and must pass) on both versions
to reduce the risk of regressions in controller functionality.