ci(*) report if a PR results in a license change
Add a new license.yaml workflow which runs on PRs. It generates a report
of dependencies and their licenses for the base branch and PR using
https://github.com/google/go-licenses

If the report differs between the two, the workflow adds a comment
showing the diff and adds a "ci/license-status-change" label to the PR.
It does not run again unless the label is manually removed after.