fix: work around a timing issue in reconcilation
A timing issue exists wherein the secret controller can reconcile
against an admin API of a Kong proxy which is currently being changed
(e.g. ENV configuration on pods modified) which can result in the
configuration in flight being lost, and reconcilation losing track
of it (because it thinks it's completed successfully).

This commit takes care of this in tests by showing the need to wait
for changes to the proxy when adding UDP/TCP routes.

I've created a follow up item to automate this away and re-apply
all configurations in the event of upgrades to the proxy:

https://github.com/Kong/kubernetes-ingress-controller/issues/1142