ci(build) add workflows to build Docker images
Add two new GitHub Actions workflows to build development and release
images.

The development images build on pushes to "main" and "next" branches,
and create "<branch>-<platform>" and "<branch>" image tags. The
"<branch>" tag is an alias for the "<branch>-alpine" tag.

The release images build for non-RC tags, and create "<tag>-<platform>",
"<tag>", and "latest" image tags. The tags without platform information
are aliases for the "<tag>-alpine" tag.