docs: add controller management to non-goals
For the Kubebuilder re-architecture we've found there are
timing and logistical problems that keep us from some of the
changes the maintainers desired regarding breaking down the
current controller monolith. This commit updates the relevant
KEP to better reflect the state of things, and we intend to
follow up on this in a later iteration/milestone.