ci(license) ignore unsuccessful label removal
Move unchanged label removal to its own step, which ignores errors.

The delete label function throws a 404 HttpError if the label is not
actually present, and for some reason wrapping it with a try/catch does
not actually suppress the error (although the catch block executes, the
step still fails). To work around this, we can use an additional step
that ignores failures.