fix(mgr) align tag flag with 1.x
Railgun originally accepted a single Kong tag only, using a flag with a
different name than we used in 1.x and a different default tag. This
aligns the flag name with 1.x, converts it to a string slice to allow
multiple tags again, and uses the same default tag set as 1.x.