fix: increase DefaultSyncSeconds
in my testing, DBLESS and Postgres mode have both shown problems with
repetitive sub-second updates. This is something we're tracking as a
potential bug elsewhere, but for now increasing the interval seems to
make the most sense for stability, as testing has shown it reduces/removes
the timing problems which we've encountered with subsecond interval updates.