fix: increase the default proxy sync seconds
It was found during testing (of DBLESS) that updating the
proxy consistently every 50ms, 500ms, or even 1s would
occasionally (and magnitudinally) increase the failure rate
of the tests as it appeared Kong would report success on
updates but the proxy state would be lost. This works around
that problem by setting a default more in line with previous
KIC 1.x work, and we will follow up on the underlying issue.