fix: cleanup object version updates in plugin tests
The previous tests would flake because the status of the
underlying Ingress resources might change concurrently
during the tests. This fix takes some time to try and
collect the latest version of the object before updating
to avoid a flake where the update would fail when an
update was attempted on a version that was too old.