fix: don't write val from failed kong update
While debugging I noticed that when an error occurs here
we would still technically assign the value provided from
the kongUpdater function to the proxy object. In practice
I did not observe this being a problem, but fundamentally
in Go we don't rely on values from functions that produce
errors and so this was a micro optimization to account
for that and improve it.