feat(mgr) negotiate Ingress versions
Add logic to check the Ingress versions available on the cluster and
enable the controller for the newest available version.

Fix the CRD version of the KongClusterPlugin detector.

Update integration tests to load Knative CRDs before starting
controller, so that the Knative detector does not disable the Knative
controller and break Knative tests.