fix: knative type errors in manager log
Previous to this patch a goroutine would be started that
would wait for knative.Ingress (CRD) to become available
and once it became available, it would load the controller
for that API dynamically. This would consistently log
errors with the type, and with these changes and moving
that functionality into a dedicated controller (rather than
a vanilla background goroutine) the logging output from
the controller is cleaner and we use controller runtime to
watch for the object updates rather than using traditional
client watchers, which makes this more consistent with the
rest of our codebase.