fix: add back the --sync-period flag
This patch adds the --sync-period flag back and delivers
the provided time duration to controller runtime, which
now in v2 handles the relevant client-go configuration on
our behalf (whereas in v1 we used it directly).

We now use the default value of 48 hours instead of the
value we previously used in v1 of 10 minutes, as per the
improvements and context relevant to the setting which make
relying on it no longer necessary.

See also: https://github.com/kubernetes-sigs/controller-runtime/issues/521