feat(status) restore publish-status-address
This reverts commit 2fae8d2674687e8dd23b512bb2130d6e6a2ce8ec.

Restore the --publish-status-address flag and associated logic in the
status updater. This flag takes a user-provided IP address and uses it
instead of attempting to determine an address from the proxy Service,
for cases where that information is not available or does not reflect
the effective external address of the proxy.