chore: remove knative auto controller enablement
This routine was previously started as we had an understanding
of the V1 implementation that it was automatically and dynamically
enabling knative support on the fly. However after re-reading the
V1 manager code, it turns out this is not the case and so having
this functionality is effectively a new feature. This patch removes
this feature so that we don't add it to the scope and maintenance
of the upcoming 2.0 release as there weren't any requirements for
it at the time, and to keep maintenance and scope tight.

We will re-asses proactive and dynamically loaded controllers in
a later iteration, see:

https://github.com/Kong/kubernetes-ingress-controller/issues/1449