chore: nonfunctional fixes for lint errors
Previously we had the linter turned off on the railgun dir
so as to be able to save a few seconds here and there and
keep things flowing quickly. Now as we are headed towards
release of 2.0 beta however, it's a good time to turn these
on. This patch fixes all the lint errors present when running
golangci-lint run ./... after enabling the railgun directory.